Extract CLI config logic into bitnet-cli-config-core
