+++ Cargo.toml
@@ -52,6 +52,7 @@ members = [
   "crates/bitnet-server-health-types-core", # SRP: reusable server health endpoint contract types
   "crates/bitnet-server",
   "crates/bitnet-cli-sampling-core", # SRP: reusable CLI sampling primitives
+  "crates/bitnet-cli-config-core", # SRP: reusable CLI configuration contracts/validation
   "crates/bitnet-scoring-core", # SRP: reusable NLL/logit scoring primitives
   "crates/bitnet-cli",
   "crates/bitnet-st2gguf",      # SafeTensors to GGUF converter
@@ -152,6 +153,7 @@ default-members = [
   "crates/bitnet-receipts-core",
   "crates/bitnet-sampling",
   "crates/bitnet-cli-sampling-core", # SRP: reusable CLI sampling primitives
+  "crates/bitnet-cli-config-core", # SRP: reusable CLI configuration contracts/validation
   "crates/bitnet-cli",
   "crates/bitnet-scoring-core",
   "crates/bitnet-server",
