+++ crates/bitnet-cli-config-core/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "bitnet-cli-config-core"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "SRP core crate for BitNet CLI configuration contracts and validation"
+
+[dependencies]
+anyhow.workspace = true
+serde = { workspace = true, features = ["derive"] }
+toml.workspace = true
+dirs.workspace = true
+tracing.workspace = true
