+++ crates/bitnet-cli/Cargo.toml
@@ -38,6 +38,7 @@ bitnet-validation = { path = "../bitnet-validation", version = "0.2.1-dev" }
 bitnet-sampling = { path = "../bitnet-sampling", version = "0.2.1-dev" }
 bitnet-eval-core = { path = "../bitnet-eval-core", version = "0.2.1-dev" }
 bitnet-cli-sampling-core = { path = "../bitnet-cli-sampling-core", version = "0.2.1-dev" }
+bitnet-cli-config-core = { path = "../bitnet-cli-config-core", version = "0.2.1-dev" }
 bitnet-scoring-core = { path = "../bitnet-scoring-core", version = "0.2.1-dev" }
 bitnet-ggml-ffi = { path = "../bitnet-ggml-ffi", version = "0.2.1-dev", optional = true }
 candle-core.workspace = true
@@ -58,7 +59,6 @@ humansize.workspace = true
 crossterm.workspace = true
 clap_complete.workspace = true
 num_cpus.workspace = true
-dirs.workspace = true
 chrono = { workspace = true, features = ["serde"] }
 memmap2.workspace = true
 sha2 = "0.10.9"
